Tidy up Sensor spec
