+++ custom_components/azrouter_integration/entity_description.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from enum import Enum
-from typing import TYPE_CHECKING, Any, override
+from typing import TYPE_CHECKING
 
 from homeassistant.components.binary_sensor import (
     BinarySensorDeviceClass,
@@ -30,81 +29,18 @@ from .api_request_composer import ApiRequestComposer, BoolToNumConverter
 from .data_value_accessor import DataValueAccessor
 from .device import AZCharger, AZDeviceBase, AZDeviceFactory, AZRouter
 from .enums import ChargeStatus
+from .sensor_specs import (
+    BinarySensorSpec,
+    EnumSensorSpec,
+    EnumValueInterpreter,
+    SensorSpec,
+    SwitchSpec,
+)
 
 if TYPE_CHECKING:
     from homeassistant.helpers.device_registry import DeviceInfo
 
     from .coordinator import AZRouterDataUpdateCoordinator
-
-
-# ── Value interpreter ─────────────────────────────────────────────────────────
-
-
-class RawValueInterpreter:
-    """Converts a raw coordinator value into a sensor native value."""
-
-    def interpret(self, raw: Any) -> Any:
-        """Return the native value for the given raw value."""
-        return raw
-
-
-class EnumValueInterpreter[EnumT: Enum](RawValueInterpreter):
-    """Interprets a raw int as the name of the matching member of EnumT."""
-
-    def __init__(self, enum_class: type[EnumT]) -> None:
-        """Store the enum class used for interpretation."""
-        self._enum_class = enum_class
-
-    def interpret(self, raw: Any) -> str | None:
-        """Return the enum member name for the given int value, or None if unknown."""
-        try:
-            return self._enum_class(raw).name
-        except (ValueError, TypeError):
-            return None
-
-
-# ── Spec dataclasses ──────────────────────────────────────────────────────────
-
-
-@dataclass
-class SensorSpec:
-    """Bundles a SensorEntityDescription with its data path and device."""
-
-    description: SensorEntityDescription
-    path: str
-    device_info: DeviceInfo
-
-    def get_value_interpreter(self) -> RawValueInterpreter | None:
-        return None
-
-
-@dataclass
-class EnumSensorSpec(SensorSpec):
-    """SensorSpec that translates a raw int into the matching enum member name."""
-
-    value_interpreter: RawValueInterpreter
-
-    def get_value_interpreter(self) -> RawValueInterpreter:
-        return self.value_interpreter
-
-
-@dataclass
-class BinarySensorSpec:
-    """Bundles a BinarySensorEntityDescription with its data path and device."""
-
-    description: BinarySensorEntityDescription
-    path: str
-    device_info: DeviceInfo
-
-
-@dataclass
-class SwitchSpec:
-    """Bundles a SwitchEntityDescription with its accessor, request, and device info."""
-
-    description: SwitchEntityDescription
-    reader: DataValueAccessor
-    writer: ApiRequestComposer
-    device_info: DeviceInfo
 
 
 @dataclass
