+++ custom_components/azrouter_integration/sensor.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 from homeassistant.components.sensor import SensorEntity
 
 from .entity import AZRouterIntegrationEntity
-from .entity_description import SensorSpec, create_entity_factory
+from .entity_description import create_entity_factory
 
 if TYPE_CHECKING:
     from homeassistant.core import HomeAssistant
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
 
     from .coordinator import AZRouterDataUpdateCoordinator
     from .data import AZRouterIntegrationConfigEntry
-    from .entity_description import RawValueInterpreter
+    from .sensor_specs import RawValueInterpreter, SensorSpec
 
 
 async def async_setup_entry(
